+++ lib/src/calendar/calendar.dart
@@ -2,6 +2,7 @@
 // Use of this source code is governed by a MIT-style license that can be
 // found in the LICENSE file.
 
+export 'actions/actions.dart';
 export 'button/button.dart';
 export 'cupertino_calendar.dart';
 export 'date_picker/date_picker.dart';