+++ lib/src/enums/calendar_dismiss_behavior.dart
@@ -4,14 +4,18 @@
 
 /// A enum that represents how the calendar can be closed.
 enum CalendarDismissBehavior {
-  /// The calendar will close when a tap occurs outside of it.
+  /// The calendar will close when a tap occurs outside of it or on the action button.
   onOutsideTap,
 
-  /// The calendar will close when a date is selected.
+  /// The calendar will close when a date is selected or on the action button.
   onDateSelect,
 
-  /// The calendar will close when either a tap occurs outside of it or a date is selected.
-  onOusideTapOrDateSelect;
+  /// The calendar will close when either a tap occurs outside of it or a date
+  /// is selected or on the action button.
+  onOusideTapOrDateSelect,
+
+  /// The calendar will close when a tap occurs on the action button only.
+  onActionTap;
 
   /// This is `true` if the behavior is either [CalendarDismissBehavior.onOutsideTap]
   /// or [CalendarDismissBehavior.onOusideTapOrDateSelect].